requirements-ai.txt
# AI and Machine Learning Dependencies
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1

# Computer Vision
opencv-python==4.8.1.78
ultralytics==8.0.200

# NLP
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Vector Database
pgvector==0.2.4

# API and Web
requests==2.31.0
httpx==0.25.2
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0

# Image Processing
imageio==2.31.5
scikit-image==0.21.0
matplotlib==3.7.2
seaborn==0.12.2

# Development
jupyter==1.0.0
ipykernel==6.25.2
matplotlib-inline==0.1.6
